requirements.txt
fastapi==0.95.1
PyYAML==6.0
pymodbus==3.2.2
